made it easier to add new resources in the future
changed the resource gathering system so that it uses an enum for resources and arrays for text objects. This means that it's much quicker to add new resources into the game.Simply add a new resource type to the resource type enum, and add a new text object to the text box array on the user interface. Everything works automatically.