Changed coordinate system, and changed procedural generation method
Coordinates are now based on the index of the star in the stars array, rather than it's position in world space.

The enviroment is generated using a simple breadth-first search algorithm, so that all the stars will always be connected together.

TODO: fix bug where stars connect to all neighbours.