Basic Expansion AI Framework
Added key AI framework. AI is devided into several managers. Each manager may have sub-managers with a return value measuring the strength of their manegement, this helps the main manager decide which sub-manager to use. For example, an Economy manager might have three ways to spend it's resources: on the military, on the colony, or on expansion. Deviding the AI into seperate classes hel;ps keep the code organised. A test has been made and the AI can already expand on it's own.