Search Algorithm overhaul
made a GraphNode into a class with functions for calculating it's value, creating it's neighbours and checking if it should be added to the frontier. Beacuse these functions are now part of a class, it makes it easy to make different graph node child classes for different purposes. While the main search algorithms can remain the same. This replaces an older system where pathNodes and other kinds of nmodes were seperate, making some unnessecery code and less ability to change.

These Search algorithms will make it easier to implement AI planning algorithms, aswell as implement spacial awareness.