Added species animations
Species will now be able to exist in the game, as part of the population tab.The current species animation can be changed with a simple integer on the animator controller.

Now that both the pie-chart and the species animations are implemented, a full species overview can be implemented ion the population tab. But first the pop-growth mechanic needs to be implemented.