Food
Population now eats food. Planet colonies are now stored in an array under the star economy script.When buildings are built, the resource production on the planet is improved, which then improves the total resource production on the star. The star updates the global resource production after it's own total resource production changes.

The structure is a little inter-connected, but it makes sense. Also fixed a number of bugs involving UI and resource-production.