Added Population Job Efficiency
If there's not enough population for every job on the planet, the planet will produce less resources. This is displayed on the UI using tool-tips and text-fields. This re-places the old idea of making population unlock building space, beacuse this method allows the user to queue several buildings before population has fully grown.