new fleet class
The old "Navigator" class that handled fleet movement and conflict has been replaced by a more easy to use, cleaner and more  modifiable version. This version of the Fleet class uses an enumeration for it's current active state, and it uses a state-machine like system to transition between them.It has an array of "orders" that can be given to it and the fleet will complete them in order. Order classes are easy to make, this allows for empires to have unique orders that they can give to their fleets. Fleets now have conflict reactions, so a fleet can stop it's current state and run-away when the fleet engages in conflict. The main reason for re-vamping the fleet system was to allow for Unique fleet mechanics which will hopefully help combat a death-ball problem within the game.