Fixed bug Preventing conflict re-actions from working.
Issue: when a Fleet enters a hostile system with the conflict--action to flee they would not flee, instead they would attempt to go to the next star in their path ignoring the problem. However, as the flee re-action changed their target coordinates, a null-reference error would occur, as the next star in the path would not exist. Fixed the bug by re-arranging code so that the path-following task gets overwritten by the escape task properly.