New A*n pathfind algorithm
Looked online, discovered that my A* pathfind algorithm wasn't implemented correctly, made a new one. This still isn't the best way to create the A* algorithm, but at least it is simple to understand. Besides the algorithm never needs many iterations anyway in the game, as it's already made from nodes that are connected across vast distances. A grid-based game would need more optimisation.