Simplified weapons code
Simplified the fleet combat. Weapons no longer use coroutines and they constantly land their targets without the ship having to aim. This effectively ensures that a battle will terminate eventually.  Weapons are turned off and on at the start and end of a battle from the fleet finite state machine.