AI abstract classes
The AI developed for the strategy game uses a hierarchical Utility AI approach. An AI is composed of AI modules, each AI module performs analysis on the current game state and passes the analysis on to multiple sub-modules.  Higher in the tree, the AI performs analysis that applies to multiple sub-modules, and at lower levels, the analysis applies to fewer and fewer sub-modules. At the very bottom of the hierarchy, individual decision-makers select appropriate actions depending upon the values derived from the process above. Each action is scored based on a set of considerations that extract relevant information from the hierarchical analysis. These considerations are in the form of scriptable objects, so designers can modify the weightings of individual considerations.