Fixed errors with HTN and created Influence Map
HTN was adding instead of inserting when undoing some actions in the plan during backtracking. It is necessary to backtrack the plan if it does not succeed. Deleted some testing HTN instances from the project folder too.  Also made an influence map class that can be used to make evaluations of the game state.