Added Influence Map Evaluation
Added the static Eval class for generating influence map propagation for fleets and territories. Fleets are evaluated by multiplying the total health by the total damage and multiplying the result by the nth triangle number, where n is the number of ships within the fleet.  The triangle number is to include the fact that fleets grow exponentially more powerful the more ships can attack at the same time, if ships attack sequentially then the total power is greatly reduced.