Find Best Path pathfinding algorithm
Using an evaluation function, findbestpath is a function that will find a path using influence maps. Each star along the path is scored based on the influence maps in the game, findbestpath will total the scores along each path and choose the path with the highest total score. It is similar to the A* algorithm except the f cost uses the evaluation instead of the heuristic, and the g cost is used solely for as a depth limit.